Organizing category_item and inserting dismissible into TaskItem
